Dodanie pobrania zamowienia po id. Refaktoring

+++ src/main/java/pl/kielce/tu/iui/iui/repository/OrderRepository.java
@@ -3,7 +3,7 @@ package pl.kielce.tu.iui.iui.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.kielce.tu.iui.iui.entity.Order;
 
-public interface OrderRepository extends JpaRepository<Order, Long>
+public interface OrderRepository extends JpaRepository<Order, Integer>
 {
 
 }
